+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Language switcher design
+- Make toasts closable
 
 ## [0.4.1] - 2025-02-22